- move the builder for ConfirmDialog instances to the ConfirmDialog class
